+++ backend/canisters/ogy_token_swap/api/src/queries/get_swap_info.rs
@@ -1,0 +1,16 @@
+use candid::CandidType;
+use ic_ledger_types::BlockIndex;
+use serde::Deserialize;
+
+use crate::token_swap::SwapInfo;
+
+#[derive(Deserialize, CandidType)]
+pub struct Args {
+    pub block_index: BlockIndex,
+}
+
+#[derive(Deserialize, CandidType)]
+pub enum Response {
+    Success(SwapInfo),
+    InternalError(String),
+}