+++ backend/canisters/ogy_token_swap/api/src/updates/request_deposit_account.rs
@@ -1,0 +1,13 @@
+use candid::{ CandidType, Principal };
+use ic_ledger_types::AccountIdentifier;
+use serde::Deserialize;
+
+#[derive(Deserialize, CandidType)]
+pub struct Args {
+    pub of: Option<Principal>,
+}
+
+#[derive(Deserialize, CandidType)]
+pub enum Response {
+    Success(AccountIdentifier),
+}