+++ backend/canisters/ogy_token_swap/impl/src/queries/mod.rs
@@ -1,4 +1,5 @@
 mod candid;
 mod http_request;
-mod request_deposit_account;
 mod get_swap_info;
+
+pub use get_swap_info::*;