+++ backend/canisters/ogy_token_swap/impl/src/updates/mod.rs
@@ -1,2 +1,6 @@
-pub mod swap_tokens;
+mod request_deposit_account;
+mod swap_tokens;
 // pub mod withdraw_deposit;
+
+pub use request_deposit_account::*;
+pub use swap_tokens::*;