+++ backend/canisters/ogy_token_swap/CHANGELOG.md
@@ -4,3 +4,6 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
+
+- first version of ogy_token_swap
+- let's users swap their legacy ogy token to new ogy tokens